+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+* [BREAKING] `Analysis#getDocument` now returns a `Result` object. When
+  `result.successful` is true, `result.value` is a Document. When
+  `result.successful` is false, then `result.value` is
+  either a Warning or undefined.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.5] - 2017-12-15